% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{check_date_format}
\alias{check_date_format}
\title{this is needed if someone opens the database in excel and saves it as a csv, the dates format changes in this case
Also NSF dates are m-d-y whereas R dates are y-m-d
potentially there is a more elegant solution than the one here
Forcing date columns to y-m-d}
\usage{
check_date_format(awards_db)
}
\description{
this is needed if someone opens the database in excel and saves it as a csv, the dates format changes in this case
Also NSF dates are m-d-y whereas R dates are y-m-d
potentially there is a more elegant solution than the one here
Forcing date columns to y-m-d
}
