% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{create_new_ticket_correspondence}
\alias{create_new_ticket_correspondence}
\title{Generalized version of the create_ticket and send_correspondences function}
\usage{
create_new_ticket_correspondence(db, template, test = TRUE)
}
\arguments{
\item{db}{(dataframe) a database containing the name and email for follow up with the following column named: first_name, id, title}

\item{template}{(character) an email template that will fill in the name and award title at each %s (see example below)}

\item{test}{(logical) sends a test email (replaces all the emails with a test email address)}
}
\value{
dataframe
}
\description{
For sending reminder emails or any other mass customized email correspondences. 
Similar to a mail merge for sending mass emails to RT.
}
\examples{
\dontrun{
db <- import_awards_db(file.path(system.file('example_db.csv', package = 'awardsBot')))
create_new_ticket_correspondence(db = db, 
                                 "Dear \%s, 
                                 \n We are writing to you today about your NSF Arctic Sciences award \%s \%s. 
                                 \n The Arctic Data Center Support Team", 
                                 test = TRUE)
}
}
