% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{create_ticket}
\alias{create_ticket}
\title{Create New Tickets}
\usage{
create_ticket(award, requestor)
}
\arguments{
\item{award}{(character) NSF award number.}

\item{requestor}{(character) PI email}
}
\value{
ticket_id (character) Newly generated RT ticket id
}
\description{
Run this code to create a new RT ticket based off NSF award number and requestor
(PI) email.  Requires a valid RT login in order to run.
}
