% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{create_ticket_and_send_initial_correspondence}
\alias{create_ticket_and_send_initial_correspondence}
\title{Create New Tickets and send initial correspondences}
\usage{
create_ticket_and_send_initial_correspondence(awards_db, database_path)
}
\arguments{
\item{awards_db}{(data.frame) database of NSF awards pulled from NSF-API}

\item{database_path}{(character) the path to save the database to}
}
\value{
awards_db (data.frame) The initial database with updated RT ticket numbers
}
\description{
Run this code to create new RT tickets and send an initial correspondence, based 
off a database of new NSF awards.  The database must include: fund_program_name,
pi_email, pi_first_name, id (NSF award #), title (NSF award title).
}
