% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_check_correspondence.R
\name{format_history_entry}
\alias{format_history_entry}
\title{Format the slack message}
\usage{
format_history_entry(msg, trunc_at = 200)
}
\arguments{
\item{msg}{(character) an rt_api object with the response}

\item{trunc_at}{(numeric) the length of the correspondence to include in the slack message}
}
\value{
(character) formatted slack message text
}
\description{
Format the slack message
}
\examples{
\dontrun{
response <- rt::rt_ticket_history_entry("21866", "520747")
format_history_entry(response)
}
}
