% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_awards.R
\name{get_awards}
\alias{get_awards}
\title{Get Awards Database}
\usage{
get_awards(database_path, n_days)
}
\arguments{
\item{database_path}{(character) path to file containing database}

\item{n_days}{(numeric) number of days to search backwards from the current day to get new awards}
}
\description{
Run this code to get the awards database with updated awards
}
