% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_check_correspondence.R
\name{get_recent_incoming_correspondence}
\alias{get_recent_incoming_correspondence}
\title{Get recent incoming correspondence formatted}
\usage{
get_recent_incoming_correspondence(ticket_id, after)
}
\arguments{
\item{ticket_id}{(character) the RT ticket number}

\item{after}{(character) check all the correspondences sent after this date}
}
\value{
list of correspondences per ticket to send
}
\description{
Get recent incoming correspondence formatted
}
