% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_awards.R
\name{import_awards_db}
\alias{import_awards_db}
\title{Read in awards}
\usage{
import_awards_db(DATABASE_PATH)
}
\arguments{
\item{DATABASE_PATH}{(character) globally defined}
}
\description{
Run this to get the awards database file
}
