% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_rt.R
\name{login_rt}
\alias{login_rt}
\title{Login to RT}
\usage{
login_rt(RT_username)
}
\arguments{
\item{RT_username}{(character) rt username}
}
\description{
Run this code to login to RT
RT Passwords are stored in keychain "r-RT"
If no passwords is stored, will ask for password and create new key
}
\details{
Note, to work with a cron job the following must be true:
key must be moved into the System keychain from the login keychain manually (within Keychain Access)
key has to have "Allow all applications to access this item" set to true
}
