% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{main function}
\usage{
main(
  database_path = Sys.getenv("DATABASE_PATH"),
  lastrun_path = Sys.getenv("LASTRUN_PATH"),
  current_date = as.character(Sys.Date()),
  annual_report_time = Sys.getenv("ANNUAL_REPORT_TIME"),
  initial_aon_offset = Sys.getenv("INITIAL_AON_OFFSET"),
  aon_recurring_interval = Sys.getenv("AON_RECURRING_INTERVAL")
)
}
\arguments{
\item{database_path}{(character) Path to the database of awards and correspondences}

\item{lastrun_path}{(character) Determines where the bot stores its state}

\item{current_date}{(character) today's date}

\item{annual_report_time}{(numeric) time in months after 'start_date' to send the first annual report reminder}

\item{initial_aon_offset}{(numeric)Number of months after award startDate to send first AON data due reminder}

\item{aon_recurring_interval}{(numeric) Number of months to send recurring emails for AON data due}
}
\value{
nothing
}
\description{
main function
}
