% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_correspondence.R
\name{send_correspondence}
\alias{send_correspondence}
\title{Send Correspondence}
\usage{
send_correspondence(adc_nsf_awards, name)
}
\arguments{
\item{adc_nsf_awards}{(data.frame) result of \code{get_awards}}

\item{name}{(character) name of person to use as email sender}
}
\description{
Run this code to send new emails
}
