% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{send_correspondence_at_time_x}
\alias{send_correspondence_at_time_x}
\title{General function that sends a correspondence based on a specified time}
\usage{
send_correspondence_at_time_x(awards_db, starting_point, direction, years = 0,
  months = 0, days = 0, rtTicket, email_text)
}
\description{
This function sends a correspondence based on a specified time interval from 
the startDate or the expDate.  You can specify which direction in time you'd like
to go based on the starting point, as well as the time interval in years, months,
and days.
}
