% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{send_correspondences}
\alias{send_correspondences}
\title{Wrapper function that sends all correspondences}
\usage{
send_correspondences(awards_db)
}
\description{
Wrapper function that sends all correspondences
}
