% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{update_aon_data_due_date}
\alias{update_aon_data_due_date}
\title{Update AON data due dates}
\usage{
update_aon_data_due_date(awards_db, aon_recurring_interval)
}
\description{
Update AON data due dates
}
