% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_awards.R
\name{update_awards}
\alias{update_awards}
\title{Update Awards Database}
\usage{
update_awards(awards_db, from_date, to_date)
}
\arguments{
\item{awards_db}{(data.frame) awards database to file containing database}

\item{from_date}{(date) date to begin search from}

\item{to_date}{(date) date to end search at}
}
\description{
Run this code to get the awards database with updated awards
}
