% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{update_contact_dates}
\alias{update_contact_dates}
\title{Update contact dates in awards database}
\usage{
update_contact_dates(awards_db, annual_report_time, initial_aon_offset,
  aon_recurring_interval)
}
\arguments{
\item{awards_db}{(data.frame) awards database}

\item{annual_report_time}{(numeric) time in months after 'start_date' to send the first annual report reminder}

\item{initial_aon_offset}{(numeric) time in months after 'start_date' to send the first aon data reminder}

\item{aon_recurring_interval}{(numeric) time in months to send aon data recurring reminders}
}
\description{
Wrapper function for set_first_annual_report_due_date, update_annual_report_due_date
set_first_aon_data_due_date, update_aon_data_due_date
}
