% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tickets.R
\name{create_ticket}
\alias{create_ticket}
\title{Create New Tickets}
\usage{
create_ticket(award, requestor)
}
\arguments{
\item{adc_nsf_awards}{(data.frame) result of \code{get_awards}}

\item{test_requestor}{(character) optional email to use for requestor for testing purposes}
}
\description{
Run this code to create new ticket in the awards database
}
