% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{create_ticket_and_send_initial_correspondence}
\alias{create_ticket_and_send_initial_correspondence}
\title{Create New Tickets and send initial correspondences}
\usage{
create_ticket_and_send_initial_correspondence(awards_db)
}
\arguments{
\item{awards_db}{(data.frame) database of NSF awards pulled from NSF-API using datamgmt::get_awards}
}
\value{
awards_db (data.frame) The initial database with updated RT ticket numbers
}
\description{
Run this code to create new RT tickets and send an initial correspondence, based 
off a database of new NSF awards.  The database must include: fundProgramName,
piEmail, piFirstName, id (NSF award #), title (NSF award title).
}
