% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_check_correspondence.R
\name{get_tickets_with_new_incoming_correspondence}
\alias{get_tickets_with_new_incoming_correspondence}
\title{Get incoming correspondences
modified from the submissions bot python code}
\usage{
get_tickets_with_new_incoming_correspondence(after)
}
\arguments{
\item{after}{date YYYY-MM-DD}
}
\value{

}
\description{
Get incoming correspondences
modified from the submissions bot python code
}
