% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{import_awards_db}
\alias{import_awards_db}
\title{Import awards database}
\usage{
import_awards_db(path)
}
\arguments{
\item{path}{(character) path to the awards database}
}
\description{
Import awards database as an all-character data.frame.
}
