% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_functions.R
\name{send_correspondences}
\alias{send_correspondences}
\title{Wrapper function that sends all correspondences}
\usage{
send_correspondences(awards_db, database_path)
}
\arguments{
\item{awards_db}{(data.frame) awards database}

\item{database_path}{(character) the path to save the database to}
}
\description{
Wrapper function for 'create_ticket_and_send_initial_correspondence',
'send_annual_report_correspondence', 'send_aon_correspondence', and
'send_one_month_remaining_correspondence'
}
