% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_functions.R
\name{update_awards}
\alias{update_awards}
\title{Update Awards Database}
\usage{
update_awards(awards_db, from_date, to_date)
}
\arguments{
\item{awards_db}{(data.frame) awards database}

\item{from_date}{(date) date to begin search from}

\item{to_date}{(date) date to end search at}
}
\description{
Update the awards database with new awards information from the NSF API 
(https://www.research.gov/common/webapi/awardapisearch-v1.htm) using 
`datamgmt::get_awards`.
}
